+++ packages/browser-extension/src/modules/ui/popup.tsx
@@ -1,5 +1,5 @@
 import React, { useCallback, useEffect, useRef, useState } from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import { WorkerClient } from "../ipc/client";
 import { getItemUrl } from "../service/ado/url";
 import { RecentChangedUpdate } from "../service/emitters/recent-manager";
@@ -268,9 +268,8 @@ export const PopupWindow: React.FC = () => {
   ) : null;
 };
 
-ReactDOM.render(
+createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
     <PopupWindow />
-  </React.StrictMode>,
-  document.getElementById("root")
+  </React.StrictMode>
 );